Langchain-chatbot-1.0/requirements.txt
langchain==0.0.208
unstructured 
selenium
Deeplake 
openai==0.27.8
psutil 
tiktoken
python-dotenv

